stamps/pyproject.toml
[tool.poetry]
name = "stampposition"
version = "1.0.0"
description = "drenowatz stamp position"
authors = ["Jonas Baumann <jone@jone.ch>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.1.78"
tqdm = "^4.66.1"
pillow = "^10.0.1"

[tool.poetry.group.lint.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --tb=short"
testpaths = ["tests"]

[tool.black]
line-length = 79
exclude = '''
(
  /(
      \.git
    | venv
  )/
)
'''

[tool.isort]
force_alphabetical_sort_within_sections = true
force_single_line = true
from_first = false
known_first_party = "app"
line_length = 79
lines_after_imports = 2
